Add Result class
This class represents the result of an operation that might fail. It
contains an optional value and an error code of type vk::Result.

To get the value of a Result, first ensure that no error occured by
calling is_error() and retrieve the value using value()

Calling value() when no value is present invokes undefined behaviour!